Allow to ignore missing objects on lookup-key updates
- It is relatively common case to attempt a lookup-key update on
an object that does not exist in the collection anymore. In
such case it is more convenient to ignore such update rather
than to try to create such object and throw an exception due to
inability to do so. That being said, updates with lookup-keys
resulting in missing objects are simply ignored.
- To be able to throw exception if needed, one should set
`update.throwOnLookupMissingObject` flag to `true`.